fix(Select): mock generated id in CheckboxSelect snapshot tests
