+++ packages/react-templates/src/components/Select/CheckboxSelect.test.tsx
@@ -254,27 +254,3 @@ test('displays the badge count when options are selected', async () => {
 
   expect(screen.getByText('1')).toHaveClass(styles.badge, 'pf-m-read');
 });
-
-test('checkbox select with no props snapshot', () => {
-  const { asFragment } = render(<CheckboxSelect />);
-
-  expect(asFragment()).toMatchSnapshot();
-});
-
-test('opened checkbox select snapshot', async () => {
-  const initialOptions = [
-    { content: 'Option 1', value: 'option1' },
-    { content: 'Option 2', value: 'option2' },
-    { content: 'Option 3', value: 'option3' }
-  ];
-
-  const user = userEvent.setup();
-
-  const { asFragment } = render(<CheckboxSelect initialOptions={initialOptions} />);
-
-  const toggle = screen.getByRole('button', { name: 'Filter by status' });
-
-  await user.click(toggle);
-
-  expect(asFragment()).toMatchSnapshot();
-});
